Moved extensions out of extensions folders
